+++ src/openlmis-table-form/tr-openlmis-invalid.directive.js
@@ -30,12 +30,14 @@
      * the $submitted state, or a openlmis-form-submit event is sent through
      * the scope.
      */
-    
+
     angular
         .module('openlmis-table-form')
         .directive('openlmisInvalid', directive);
 
-    function directive() {
+    directive.$inject = ['$rootScope'];
+
+    function directive($rootScope) {
         return {
             link: link,
             restrict: 'A',
@@ -47,26 +49,29 @@
                 openlmisInvalidCtrl = ctrls[0],
                 tr = element.parents('tr:first');
 
-            if(tr.length === 0) {
+            if (tr.length === 0) {
                 return;
             }
 
             openlmisInvalidCtrl.hide();
 
-            scope.$watch(function(){
+            scope.$watch(function() {
                 return tr.hasClass('is-focused');
             }, function(isFocused) {
-                if(isFocused) {
+                if (isFocused) {
                     wasFocused = true;
                 }
-                if(!isFocused && wasFocused){
-                    openlmisInvalidCtrl.show();
+                if (!isFocused && wasFocused) {
+                    displayRowErrors();
                 }
             });
 
-            scope.$on('openlmis-form-submit', function(){
+            scope.$on('openlmis-form-submit', displayRowErrors);
+
+            function displayRowErrors() {
                 openlmisInvalidCtrl.show();
-            });
+                $rootScope.$emit('openlmis.displayRowErrors');
+            }
         }
     }